chess_arena/requirements.txt
chess>=1.10.0
anthropic>=0.25.0
openai>=1.30.0
python-dotenv>=1.0.0
